Moved functionality for adding catalogue numbers to separate module
